requirements.txt
 pandas==2.2.2
 numpy==1.26.4
 scikit-learn==1.4.2
 joblib==1.4.2
 psycopg2-binary==2.9.9
 fastapi==0.111.0
 uvicorn[standard]==0.30.1
 pydantic==2.7.4
 python-dotenv==1.0.1
 streamlit>=1.33
 plotly>=5.20
 requests>=2.31

 